+++ CHANGELOG.md
@@ -22,6 +22,7 @@ omnibus-gitlab repository.
 - Remove update_all_mirrors_worker_cron and update_all_remote_mirrors_worker_cron settings 49706b
 - Expose max_standby_archive_delay and max_standby_streaming_delay Postgresql settings
 - Disconnect and reconnect database connections when forking in Unicorn a7b35aaf
+- Add support for the PostgreSQL max_replication_slots setting
 
 8.16.5